+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   within the broker.
 * Add method to expose the `BrokerInstance` in `OpenMQContainer`.
 * Add method to expose the `ConnectionFactory` field in `OpenMQContainer`.
+* Add method to create queue and topic references in `OpenMQContainer`.
 
 ## 0.82:
 * Add helper methods `AbstractServerTest.ctran` that clear the EntityManager before running block.